+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## v3.2.3
 ### Bug Fixes
 * Fix crash when a material name has slashes or other characters that are not allowed in file names.
+* Fix crash in unity 2022 when an animator is not humanoid.
 
 ## v3.2.2
 ### Bug Fixes