+++ CHANGELOG.md
@@ -1,6 +1,7 @@
-## v3.2.3
+## v3.3.0
 ### Changes
 * No longer put the dummy animation clip in empty states. [(more)](https://github.com/d4rkc0d3r/d4rkAvatarOptimizer/issues/63)
+* Add debug list `Mesh Bone Weight Stats`.
 
 ### Bug Fixes
 * Fix crash when a material name has slashes or other characters that are not allowed in file names.